Update Restaurant.cpp
All the for loops are comparing an assigned variable to a vector operator, which naturally is defined as an unsigned integer. I changed all the iterator integers in the loops to unsigned integers, so that it will work on systems (like some linux based) that cannot interchange the two. I also found that the pointer *s in the Restaurant::print_shortest_distance(queueVertex*, string, string) function was neither used nor needed, so I commented it out.